+++ plugin-maven/CHANGES.md
@@ -5,7 +5,7 @@ We adhere to the [keepachangelog](https://keepachangelog.com/en/1.0.0/) format (
 ## [Unreleased]
 
 ### Added
-- Add support for AsciiDoc formatting via `adocfmt`. (#TODO-PR-LINK)
+- Add support for AsciiDoc formatting via `adocfmt`. ([#2960](https://github.com/diffplug/spotless/pull/2960))
 
 ## [3.6.0] - 2026-05-27
 ### Added